// frontend/src/models/Organizer.ts
type Organizer = {
  _id: string;
  name: string;
  email: string;
  phone: string;
  company: string;
  is_verified: boolean;
  is_bank_account: boolean;
  bank_account_name: string;
  bank_account_number: string;
  bank_ifsc_code: string;
  bank_name: string;
  created_at: string;
  updated_at: string;
}

export default Organizer;